+++ examples/build.sh
@@ -7,18 +7,12 @@ root="$(dirname "$0")/.."
 example_root="$(dirname "$0")/$name"
 did_file="/tmp/a.did"
 
-# This script generates the did file, build the project (passed as $1) and then run the ic-wasm to shrink and attach metadata.
-cargo build --manifest-path="$example_root/Cargo.toml" \
-    --target wasm32-unknown-unknown \
-    --release \
-    --package "$package" --features "ic-cdk/wasi"
-
-wasmtime "$example_root/target/wasm32-unknown-unknown/release/$package.wasm" > $did_file
-
 cargo build --manifest-path="$example_root/Cargo.toml" \
     --target wasm32-unknown-unknown \
     --release \
     --package "$package"
+
+candid-extractor "$example_root/target/wasm32-unknown-unknown/release/$package.wasm" 2>/dev/null > $did_file || true
 
 ic-wasm "$example_root/target/wasm32-unknown-unknown/release/$package.wasm" \
     -o "$example_root/target/wasm32-unknown-unknown/release/$package.wasm" \
@@ -27,4 +21,3 @@ ic-wasm "$example_root/target/wasm32-unknown-unknown/release/$package.wasm" \
 ic-wasm "$example_root/target/wasm32-unknown-unknown/release/$package.wasm" \
     -o "$example_root/target/wasm32-unknown-unknown/release/$package-opt.wasm" \
     shrink
-