+++ examples/counter/src/inter2_rs/build.rs
@@ -4,8 +4,8 @@ use std::path::PathBuf;
 fn main() {
     let manifest_dir =
         PathBuf::from(std::env::var("CARGO_MANIFEST_DIR").expect("Cannot find manifest dir"));
-    let counter = Config::new("inter_mo");
+    let inter_mo = Config::new("inter_mo");
     let mut builder = Builder::new();
-    builder.add(counter);
+    builder.add(inter_mo);
     builder.build(Some(manifest_dir.join("declarations")));
 }