+++ src/ic-cdk-timers/CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.5.0] - 2023-09-18
+
+### Changed
+
+- Upgrade `ic-cdk` to v0.11.
+
 ## [0.4.0] - 2023-07-13
 
 ### Changed