+++ src/ic-cdk/CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.11.0] - 2023-09-18
+
+### Changed
+
+- Candid Export workflow is changed. (#424)
+  * No need to compile for WASI separately.
+  * Canisters should still invoke `ic_cdk::export_candid!()` to export candid.
+  * Then use [`candid-extractor`](../candid-extractor/) to extract candid from the canister WASM.
+
 ## [0.10.0] - 2023-07-13
 
 ### Changed